requirements.txt
absl-py==0.11.0
scipy>=1.5.4  
tensorflow==2.4.0
tensorboardX>=2.1
torch==2.0.1
torchvision==0.15.2
tqdm>=4.55.1
packaging
matplotlib
wandb